[Refactor] Issue List Page
Move issueContext and issueListReducer from IssueList to IssueListPage
