+++ front/src/components/IssueContent.js
@@ -5,7 +5,7 @@ import ClosedSvg from '../svgs/ClosedSvg';
 import OpenedSvg from '../svgs/OpenedSvg';
 import IssueInfo from './IssueInfo';
 import IssueLabel from './IssueLabel';
-import { IssueContext } from './IssueList';
+import { IssueContext } from './../pages/IssueListPage';
 
 const OutDiv = styled.div`
   display: flex;
