+++ front/src/components/IssueList.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useReducer, createContext } from 'react';
+import React, { useContext } from 'react';
 import styled from 'styled-components';
 import IssueContent from './IssueContent';
 import SelectedCount from './SelectedCount';
@@ -7,7 +7,7 @@ import AssigneeFilterButton from './AssigneeFilterButton';
 import LabelFilterButton from './LabelFilterButton';
 import MilestoneFilterButton from './MilestoneFilterButton';
 import MarkAs from './MarkAs';
-import issueAPI from './../apis/issue.api';
+import { IssueContext } from './../pages/IssueListPage';
 
 const IssueSection = styled.section`
   border: 1px solid #e1e4e8;
@@ -40,39 +40,8 @@ const RightFloatDiv = styled.div`
   width: ${(props) => props.width || 'auto'};
 `;
 
-const issueListReducer = (issueList, { type, payload }) => {
-  switch (type) {
-    case 'setInitial':
-      return payload;
-    case 'checkAll':
-      return issueList.map((issue) => {
-        return { ...issue, isChecked: payload.isChecked };
-      });
-    case 'check':
-      return issueList.map((issue) => {
-        return issue.id === payload.id
-          ? { ...issue, isChecked: !issue.isChecked }
-          : issue;
-      });
-  }
-};
-
-export const IssueContext = createContext();
-
 const IssueList = () => {
-  const [issueList, issueListDispatch] = useReducer(issueListReducer, []);
-
-  useEffect(async () => {
-    const issues = await issueAPI.getIssues();
-
-    const orderedIssues = issues
-      .map((issue) => {
-        return { ...issue, isChecked: false };
-      })
-      .sort((a, b) => (a.id > b.id ? -1 : 1));
-
-    issueListDispatch({ type: 'setInitial', payload: orderedIssues });
-  }, []);
+  const { issueList, issueListDispatch } = useContext(IssueContext);
 
   const getCountChecked = () => {
     return issueList.reduce((count, issue) => {
@@ -97,41 +66,39 @@ const IssueList = () => {
   );
 
   return (
-    <IssueContext.Provider value={{ issueList, issueListDispatch }}>
-      <IssueSection>
-        <Header>
-          <AllCheckDiv>
-            <input
-              type="checkbox"
-              name="all-issue"
-              value="all"
-              onChange={(e) =>
-                issueListDispatch({
-                  type: 'checkAll',
-                  payload: { isChecked: e.target.checked },
-                })
-              }
-            />
-          </AllCheckDiv>
-          <ExtraHeader>
-            {!(getCountChecked() > 0) ? (
-              <>{dropDownTags}</>
-            ) : (
-              <>
-                <SelectedCount count={getCountChecked()} /> {markAs}{' '}
-              </>
-            )}
-          </ExtraHeader>
-        </Header>
-        <section>
-          {issueList.map((issue, index) => {
-            return (
-              <IssueContent key={`issue${issue.id}-${index}`} data={issue} />
-            );
-          })}
-        </section>
-      </IssueSection>
-    </IssueContext.Provider>
+    <IssueSection>
+      <Header>
+        <AllCheckDiv>
+          <input
+            type="checkbox"
+            name="all-issue"
+            value="all"
+            onChange={(e) =>
+              issueListDispatch({
+                type: 'checkAll',
+                payload: { isChecked: e.target.checked },
+              })
+            }
+          />
+        </AllCheckDiv>
+        <ExtraHeader>
+          {!(getCountChecked() > 0) ? (
+            <>{dropDownTags}</>
+          ) : (
+            <>
+              <SelectedCount count={getCountChecked()} /> {markAs}{' '}
+            </>
+          )}
+        </ExtraHeader>
+      </Header>
+      <section>
+        {issueList.map((issue, index) => {
+          return (
+            <IssueContent key={`issue${issue.id}-${index}`} data={issue} />
+          );
+        })}
+      </section>
+    </IssueSection>
   );
 };
